jwt 로그인 권한 테이블 수정

+++ src/main/java/com/my/toyprj/auth/dto/MemberAuthDTO.java
@@ -26,7 +26,7 @@ public class MemberAuthDTO {
         this.auth=auth;
     }
 
-    public MemberAuth toEntity(int userNo, String auth){
+    public MemberAuth toEntity(Long userNo, String auth){
         return new MemberAuth(userNo,auth);
     }
 }
